requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
requests>=2.31.0
pymongo>=4.5.0
loguru>=0.7.0
pydantic>=2.0.0
aiohttp>=3.8.5
httpx>=0.24.1
python-multipart>=0.0.6
PyPDF2>=3.0.1
python-docx>=1.1.2
dnspython>=2.6.0
colorama>=0.4.6 